+++ hata/discord/embed/embed/embed.py
@@ -1340,3 +1340,46 @@ class Embed(RichAttributeErrorBaseType):
         """
         self.thumbnail = EmbedThumbnail(url)
         return self
+    
+    
+    @property
+    def _data(self):
+        """
+        Deprecated and will be removed in 2023 August.
+        """
+        warnings.warn(
+            (
+                f'`{self.__class__.__name__}._data` is deprecated, and will be removed in 2023 August. '
+                f'Please use `.to_data()` instead.'
+            ),
+            FutureWarning,
+            stacklevel = 2,
+        )
+        
+        return self.to_data(include_internals = True)
+    
+    
+    @_data.setter
+    def _data(self, data):
+        warnings.warn(
+            (
+                f'`{self.__class__.__name__}._data` is deprecated, and will be removed in 2023 August. '
+                f'Please use `.from_data(data)` instead.'
+            ),
+            FutureWarning,
+            stacklevel = 2,
+        )
+        
+        self.author = parse_author(data)
+        self.color = parse_color(data)
+        self.description = parse_description(data)
+        self.fields = parse_fields(data)
+        self.footer = parse_footer(data)
+        self.image = parse_image(data)
+        self.provider = parse_provider(data)
+        self.thumbnail = parse_thumbnail(data)
+        self.timestamp = parse_timestamp(data)
+        self.title = parse_title(data)
+        self.type = parse_type(data)
+        self.url = parse_url(data)
+        self.video = parse_video(data)